greatly speed up query execution
